Consolidate stable maths tests.
